Patch for compatibility with Pythonista interpreter on iOS
Until a connection is made to the internet, in Pythonista there is no way to know the IP address / interface that will be used for a connection. As such, we need to attempt to reach at least one internet site first, then get the IP address of the interface that was used for the connection.